feat: 최대 2개까지 선택 가능한 섹션 옆 경고 문구 추가
